requirements.txt
biopython==1.84
numpy==1.26.3
pandas==2.2.3
scikit-learn==1.5.2
scipy==1.14.1
torch==2.1.2
tqdm==4.67.1
einops==0.8.1
multimolecule==0.0.5
transformers==4.37.1